restructured audio logic into two main parts: 'core' and 'modules'
