+++ src/audio/modules/generator/oscillator/melodic/simple-oscillator.ts
@@ -1,8 +1,8 @@
-import { Settings } from "../../../core/settings";
+import { Settings } from "../../../../core/settings";
 import { NoteSettings } from "../../../settings/note-settings";
 
-import { RestartableGenerator, type EndableNode } from "../../../core/emitter";
-import type { FrequencyBasedSignal, PulseBasedSignal } from "../../../core/signal";
+import { RestartableGenerator, type EndableNode } from "../../../../core/emitter";
+import type { FrequencyBasedSignal, PulseBasedSignal } from "../../../../core/signal";
 
 import { Logger } from "tslog";
 import type { ILogObj } from "tslog";
