+++ src/audio/modules/intermediate-emitter/mixer/additive-mixer.ts
@@ -1,5 +1,5 @@
-import { Settings } from "../../core/settings";
-import type { Emitter } from "../../core/emitter";
+import { Settings } from "../../../core/settings";
+import type { Emitter } from "../../../core/emitter";
 
 import { Logger } from "tslog";
 import type { ILogObj } from "tslog";
