+++ src/audio/modules/settings/note-settings.ts
