+++ src/components/control/InvisibleKeyboard.tsx
@@ -3,8 +3,8 @@ import React, { useEffect, useRef, useState } from "react";
 import styles from "./InvisibleKeyboard.module.css";
 
 import { Settings } from "../../audio/core/settings";
-import { Note12TET } from "../../audio/note/note12tet";
-import { TestPolySynth, testPolySynth } from "../../audio/synth/virtual-analog/test-synth";
+import { Note12TET } from "../../audio/modules/note/note12tet";
+import { TestPolySynth, testPolySynth } from "../../audio/modules/synth/virtual-analog/test-synth";
 // import { monoSynth } from "../../audio/synth/virtual-analog/analog-synth";
 
 import { Logger } from "tslog";
